+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 - 	Changed the description of the panel (step by step explanation)
 - 	Added a panel to show account email information
 - 	Added JetRails tab containing two factor authentication option, this will show both forms
+
+## Version 1.0.2
+- 	Auto generate new secret on each page load, until validated
+- 	Removed "secret" box
+- 	Removed "Generate New Secret" button
+- 	Removed "Generate New Secret" action from configuration controller
+- 	Made description a dynamically loaded block, depending on configuration status
+- 	Updated Two Factor Authentication tab to dynamically load the description
+- 	Updated two factor stylesheet